pyproject.toml
[project]
name = "clip-eval-runner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "clip-benchmark>=1.6.2",
    "huggingface-hub[cli]>=0.34.3",
    "open-clip-torch>=3.0.0",
    "polars>=1.31.0",
    "pycocotools>=2.0.10",
    "skypilot[aws]>=0.10.0",
    "pyyaml>=6.0",
    "awscli>=1.41.16",
    "boto3>=1.39.16",
]

[project.scripts]
clipeval = "clip_eval_runner.cli:main"

[build-system]
requires = [
    "setuptools",
    "wheel"
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["clip_eval_runner"]

[dependency-groups]
dev = [
    "ruff>=0.12.5",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
